Very basic inversion function is working now.
Because of circulatity (similar to what happened with terms) I moved some of the functions our of translation and into metaterm.

Translation of abstractions is not correct now as they will need to determine LF type of the variable.

Currently there are excessive amounts of parenthesis but this can be changed in future.S