Getting unit tests to build and run
Being able to run the existing unit test battery is important to
help keep Abella correct as new developments accumulate. Changes in
the project structure have not been reflected in the test suite; a
minimal changeset to build and run these conveniently is made and
a Makefile rule put together.

Many of the existing test modules fail to compile and must be
inspected separately. Of those that do without changes, some
contain regressions which must also be addressed.