A couple of fixes to build the test suite
Make the test suite depend on the main Abella executable to solve
quickly some caching problems in _build and issues to fail to
locate, or locate twice, the Unix library.

Moreover, Test_helper itself requires part of the grammar in the
parser to be exposed as a starting symbol (or it is, at least, an
easy, if hideous, fix to the conundrum we find ourselves in).