Fix search block in tactics tests
The search tactic has changed its interface and is brought up to date in
the test suite. Namely, ~alldefs is replaced by ~def_unfold, and ~retype
is added. The fixes are the simplest within reason: ~def_unfold becomes
essentially ~alldefs, with its parameter being vacuous, and ~retype is
the simplest conversion possible, without a substitution list (a
constant term like True is obviously a "possibility").

The current changeset allows the search block to compile and all its
tests to pass, which goes to say that none of those depend on the
refinements introduced by the newer code. So basically we've gone from
red to green, and proceed with refactoring we need to add tests that
exercise these differences and drive their refinement as needed. TODO
markers signal these sinful points.