Non iterative use of search depth bound
This commit modifies the internal behavior of the search tactic (while
preserving its semantics) by moving from IDDFS on the range of dephts
from 1 to the effective maximum depth to a single execution of
Tactics.search on said maximum depth.

The rationale for this change is the following. If an explicit bound is
not given, the system falls back to the (small) default, and any penalty
incurred by trying to explore deeper and initially missing shallow
solutions is negligible.

If, on the other hand, an explicit bound is given, the responsibility
for its adequacy falls on the user, as do any adverse effects of
excessive slack on runtime. On the upside, this strategy avoids the
quadratic time factor that results from multiple executions of the
search tactic with increasing bounds, which becomes problematic for
higher values of these.